python/OPCUA2SQL/requirements.txt
opcua
mysql-connector-python